Create Equal Elements
First, we are taking input. The main logical Part of this Solution is that if we find the frequency of every element which is present in the array and store it in the dictionary named mydict after that, we find the max value from the dictionary, and after finding that we subtract it from the size of an array which is n.