Update Testing Robot
Firstly, we created a list with the initial position and then started a for loop to iterate over the list containing the characters of the string. If robot moves 'right' we increase the most recent position of the robot by 1 else it decreases by 1.

Then we measure the length of the "new" list by converting it to a set (as converting in a set removes duplicated elements).