Update Testing Robot
for lines 5 and 6, You have to take a blank array in which the values which will be updated later on, could be kept.
for lines 7 to 14, a for loop is introduced in which the values at every index in the for loop will be checked. then if the value of s[i], i.e. the value on i-th index is 'R' increase the given X value and append it in the 'arr' array. and Lastly, print the length of set values that occurred in the 'arr' array.